+++ src/effects/spells/guidance.js
@@ -28,7 +28,11 @@ export function guidanceEffect(document) {
       priority: "20",
     },
   );
-  foundry.utils.setProperty(effect, "flags.dae.specialDuration", ["isSkill", "isCheck", "isInitiative"]);
+  foundry.utils.setProperty(effect, "flags.dae.specialDuration", [
+    // "isSkill",
+    // "isCheck",
+    "isInitiative",
+  ]);
 
   document.effects.push(effect);